dev/requirements.txt
openai
google-genai
json
jsonschema
matplotlib
numpy
networkx
z3-solver
lark
sentence-transformers